refactoring for more agile App, started by organizing Tourist Info Cards and splitting into two components
