+++ src/App.js
@@ -1,13 +1,8 @@
 import React from "react";
-import glasgow from "./images/glasgow.jpeg";
-
-// Images
-import manchester from "./images/manchester.jpeg";
-import london from "./images/london.jpeg";
 
 // Components:
 import Heading from "./components/Heading";
-import TouristInfoCards from "./components/TouristInfoCards";
+import TouristInfoCards from "./components/TouristInfoCards/TouristInfoCards";
 import Footer from "./components/Footer";
 import Restaurant from "./components/Restaurant/Restaurant";
 
@@ -24,26 +19,7 @@ const App = () => {
       <Restaurant />
       <br />
       <br />
-      <div className="cities">
-        <TouristInfoCards
-          imageUrl={glasgow}
-          cityName="Glasgow"
-          description="Lovely City"
-          link="https://peoplemakeglasgow.com/"
-        />
-        <TouristInfoCards
-          imageUrl={manchester}
-          cityName="Manchester"
-          description="Lovely City Too"
-          link="https://www.visitmanchester.com/"
-        />
-        <TouristInfoCards
-          imageUrl={london}
-          cityName="London"
-          description="Lovely City Too Too"
-          link="https://visitlondon.com/"
-        />
-      </div>
+      <TouristInfoCards />
       <br />
       <br />
       <Footer
