+++ src/tourist-info.json
@@ -1,0 +1,20 @@
+[
+  {
+    "imageUrl": "https://images.unsplash.com/photo-1600209242823-7ca6899dc77d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1374&q=80",
+    "cityName": "Glasgow",
+    "description": "The ideal shopping city. From independent stores to high street retailers, Glasgow is the best shopping city. Glasgow's city centre is home to flagship stores, impressive shopping centres and designer favorites all within an easily walkable area. The city is also rich in artists and creators which makes for a thriving and exciting independent shopping scene.",
+    "link": "https://peoplemakeglasgow.com"
+  },
+  {
+    "imageUrl": "https://images.unsplash.com/photo-1560352642-d19964cf73ac?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1074&q=80",
+    "cityName": "Manchester",
+    "description": "Youthful, diverse, energetic and bursting with character; Manchester is one of the most exciting places to visit in the UK right now where everybody and anybody is very warmly welcomed. Known throughout the world as the birthplace of the industrial revolution, Manchester has a proud history in science, politics, music, arts and sport. And today the city combines this heritage with a progressive vision to be a city that delivers surprise and delight in equal measures.",
+    "link": "https://www.visitmanchester.com/"
+  },
+  {
+    "imageUrl": "https://images.unsplash.com/photo-1513635269975-59663e0ac1ad?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
+    "cityName": "London",
+    "description": "Find things to do in London, from iconic sightseeing spots and fun-filled days out to top restaurants, theatre and unmissable London events. If you’re not able to visit just yet, plan ahead to make the most of your next visit.",
+    "link": "https://visitlondon.com/"
+  }
+]
